Add files via upload
•Introduction
Information security is very vital for each organization as they operate confidential data.
Every business sector is now focusing more on how to keep their data safe from every bad aspect. Nowadays, educational institutes are focusing more on the information security framework after the cber-attack on the blackboard back in 2020. The University of Higher Education (UHE) wants us to develop an information security policies framework for their Information security management system (ISMS). This is the group activity in which every member has a specific role in the development of this framework.
In this report, we will describe the purpose and scope of these policies and will mention different roles and their responsibilities in order to implement these policies in a clear way.
In addition to that, different monitoring mechanisms will also be discussed to check and balance the information security policies to avoid any security vulnerabilities.